Move parser suite test helpers to a separate package
When parserTestSuite was added, it introduced importing the `testing`
package (via testify) outside of a test context. That caused the
`test.v` flag to be defined, so initializing the cache in normal
operations (running `wash server`) failed.

Move the parser suite test helpers to their own package that's only used
during testing to avoid this problem.
