+++ cmd/internal/find/primary/meta/numericPredicate_test.go
@@ -3,13 +3,14 @@ package meta
 import (
 	"testing"
 
-	"github.com/puppetlabs/wash/cmd/internal/find/primary/numeric"
+	"github.com/puppetlabs/wash/cmd/internal/find/parser/parsertest"
 	"github.com/puppetlabs/wash/cmd/internal/find/parser/predicate"
+	"github.com/puppetlabs/wash/cmd/internal/find/primary/numeric"
 	"github.com/stretchr/testify/suite"
 )
 
 type NumericPredicateTestSuite struct {
-	predicate.ParserTestSuite
+	parsertest.Suite
 }
 
 func (s *NumericPredicateTestSuite) TestErrors() {
